Move code to utils file
